+++ docker-compose.yml
@@ -9,7 +9,7 @@ services:
         extra_hosts:
             - "host.docker.internal:host-gateway"
         ports:
-            - "${APP_PORT:-8080}:80"
+            - "${APP_PORT:-8000}:80"
             - "${VITE_PORT:-5173}:${VITE_PORT:-5173}"
         environment:
             WWWUSER: "${WWWUSER}"
